+++ src/controllers/contractController.js
@@ -2,8 +2,6 @@ const Boom = require('boom');
 const flat = require('flat');
 const crypto = require('crypto');
 const Contract = require('../models/Contract');
-const User = require('../models/User');
-const Customer = require('../models/Customer');
 const ESign = require('../models/ESign');
 const translate = require('../helpers/translate');
 const ContractHelper = require('../helpers/contracts');
