+++ src/helpers/fundings.js
@@ -1,13 +1,10 @@
 const Boom = require('boom');
 const moment = require('moment');
-const get = require('lodash/get');
 const has = require('lodash/has');
 const Customer = require('../models/Customer');
 const { populateService } = require('./subscriptions');
 const UtilsHelper = require('./utils');
 const translate = require('./translate');
-const { DAYS_INDEX, FUNDING_FREQUENCIES, FUNDING_NATURES, CIVILITY_LIST } = require('./constants');
-const CustomerRepository = require('../repositories/CustomerRepository');
 
 const { language } = translate;
 
@@ -54,77 +51,6 @@ exports.populateFunding = (funding, subscriptions) => {
   else funding.subscription = { ...sub };
 
   return funding;
-};
-
-const fundingExportHeader = [
-  'Titre',
-  'Nom',
-  'Prénom',
-  'Tiers payeur',
-  'Nature',
-  'Service',
-  'Date de début',
-  'Date de fin',
-  'Numéro de dossier',
-  'Fréquence',
-  'Montant TTC',
-  'Montant unitaire TTC',
-  'Nombre d\'heures',
-  'Jours',
-  'Participation du bénéficiaire',
-];
-
-exports.exportFundings = async (credentials) => {
-  const customerFundings = await CustomerRepository.getCustomerFundings(get(credentials, 'company._id', null));
-  const data = [fundingExportHeader];
-
-  for (const cus of customerFundings) {
-    const fundInfo = [];
-    if (cus.identity) {
-      fundInfo.push(
-        CIVILITY_LIST[get(cus, 'identity.title')] || '',
-        get(cus, 'identity.lastname', '').toUpperCase(),
-        get(cus, 'identity.firstname', '')
-      );
-    } else fundInfo.push('', '', '');
-
-    let { funding } = cus;
-    if (!funding) fundInfo.push('', '', '', '', '', '', '', '', '', '', '', '');
-    else {
-      funding = UtilsHelper.mergeLastVersionWithBaseObject(funding, 'createdAt');
-
-      const nature = FUNDING_NATURES.find(nat => nat.value === funding.nature);
-      const lastServiceVersion = has(funding, 'subscription.service.versions')
-        ? UtilsHelper.getLastVersion(funding.subscription.service.versions, 'startDate')
-        : null;
-      const frequency = FUNDING_FREQUENCIES.find(freq => freq.value === funding.frequency);
-      let careDays = '';
-      if (funding.careDays) {
-        funding.careDays.map((dayIndex) => {
-          careDays = careDays.concat(`${DAYS_INDEX[dayIndex]} `);
-        });
-      }
-
-      fundInfo.push(
-        funding.thirdPartyPayer ? (funding.thirdPartyPayer.name || '') : '',
-        nature ? nature.label : '',
-        lastServiceVersion ? lastServiceVersion.name : '',
-        funding.startDate ? moment(funding.startDate).format('DD/MM/YYYY') : '',
-        funding.endDate ? moment(funding.endDate).format('DD/MM/YYYY') : '',
-        funding.folderNumber || '',
-        frequency ? frequency.label : '',
-        UtilsHelper.formatFloatForExport(funding.amountTTC),
-        UtilsHelper.formatFloatForExport(funding.unitTTCRate),
-        UtilsHelper.formatFloatForExport(funding.careHours),
-        careDays || '',
-        UtilsHelper.formatFloatForExport(funding.customerParticipationRate)
-      );
-    }
-
-    data.push(fundInfo);
-  }
-
-  return data;
 };
 
 exports.createFunding = async (customerId, payload) => {
