+++ src/helpers/subscriptions.js
@@ -8,7 +8,6 @@ const isEqual = require('lodash/isEqual');
 const Customer = require('../models/Customer');
 const translate = require('../helpers/translate');
 const UtilsHelper = require('../helpers/utils');
-const { CIVILITY_LIST } = require('./constants');
 
 const { language } = translate;
 
@@ -58,56 +57,6 @@ exports.subscriptionsAccepted = (customer) => {
     }
   }
   return customer;
-};
-
-const subscriptionExportHeader = [
-  'Titre',
-  'Nom',
-  'Prénom',
-  'Service',
-  'Prix unitaire TTC',
-  'Volume hebdomadaire estimatif',
-  'Dont soirées',
-  'Dont dimanches',
-];
-
-exports.exportSubscriptions = async (credentials) => {
-  const customers = await Customer
-    .find({ subscriptions: { $exists: true, $not: { $size: 0 } }, company: get(credentials, 'company._id', null) })
-    .populate({ path: 'subscriptions.service' })
-    .lean();
-  const data = [subscriptionExportHeader];
-
-  for (const cus of customers) {
-    for (const sub of cus.subscriptions) {
-      const subInfo = [];
-      if (cus.identity) {
-        subInfo.push(
-          CIVILITY_LIST[get(cus, 'identity.title')] || '',
-          get(cus, 'identity.lastname', '').toUpperCase(),
-          get(cus, 'identity.firstname', '')
-        );
-      } else subInfo.push('', '', '');
-
-      const lastServiceVersion = UtilsHelper.getLastVersion(sub.service.versions, 'startDate');
-      if (lastServiceVersion) subInfo.push(lastServiceVersion.name);
-      else subInfo.push('');
-
-      const lastVersion = UtilsHelper.getLastVersion(sub.versions, 'createdAt');
-      if (lastVersion) {
-        subInfo.push(
-          UtilsHelper.formatFloatForExport(lastVersion.unitTTCRate),
-          UtilsHelper.formatFloatForExport(lastVersion.estimatedWeeklyVolume),
-          lastVersion.evenings || '',
-          lastVersion.sundays || ''
-        );
-      } else subInfo.push('', '', '', '');
-
-      data.push(subInfo);
-    }
-  }
-
-  return data;
 };
 
 exports.updateSubscription = async (params, payload) => {
